Add sample weights, Newton-weighted leaves, multiclass + cat support, sklearn integration (#1)
Adds sklearn-style sample_weight to fit on DecisionTree, RandomForestClassifier,
RandomForestRegressor, RFGBoostClassifier, and RFGBoostRegressor. Weights flow
through to weighted Gini, weighted variance, and weighted leaf means in tree.rs;
class_counts changes from HashMap<usize,usize> to HashMap<usize,f64> for weighted
counts. Bootstrap remains uniform; weights apply at split-time and at leaves
(sklearn convention).

Switches the boosting classifier's leaf computation from FHT2000 working response
mean to Newton-weighted leaves: target = (y-p)/h, leaf weight = sample_weight * h
where h = p(1-p). The weighted leaf mean reduces to Sum(g)/Sum(h), the optimal
Newton step, matching XGBoost's loss-second-order treatment exactly.

Wires the multiclass + categorical path: when n_classes > 2 and cat_features is
set, the wrapper uses fit_matrix_multiclass / transform_matrix_multiclass for
WOE encoding (one column per (feature, class) pair). Multiclass softmax in Rust
already existed; only the Python wrapper was binary-only.

Wraps the Python classes with sklearn BaseEstimator + ClassifierMixin /
RegressorMixin so RFGBoost works in Pipeline, GridSearchCV, cross_val_score
(MRO order is mixin first to chain __sklearn_tags__ correctly on sklearn 1.6+).

Renames Rust pyclass RandomForest -> RandomForestClassifier and drops the
Python alias indirection.

Other:
- validate_finite NaN/Inf guard at fit boundaries
- validate_weights non-negative finite check
- random_state=None now uses OS entropy (sklearn convention); seed_rng helper
- Default n_estimators=20, rf_n_estimators=20

Empirical (credit risk, 20k train, WOE encoding):
- RFGBoost+WOE Gini 0.8665 -> 0.8677, Brier 0.05006 -> 0.04984
- Best Brier among XGBoost (0.05124), LightGBM (0.05101), CatBoost (0.04994)

Tests: 16 new (5 Newton equivalence + 11 sample_weight + 4 multiclass + cat),
all closed-form algebraic checks pass to ~1e-9 precision.

+++ src/tree_shap.rs
@@ -36,10 +36,10 @@ impl TreeSHAP {
                 let value = if self.model_type == "classification" {
                     let mut probs = vec![0.0; self.n_classes];
                     if let Some(counts) = &n.class_counts {
-                        let total: usize = counts.values().sum();
-                        if total > 0 {
+                        let total: f64 = counts.values().sum();
+                        if total > 0.0 {
                             for (&cls, &cnt) in counts {
-                                if cls < self.n_classes { probs[cls] = cnt as f64 / total as f64; }
+                                if cls < self.n_classes { probs[cls] = cnt / total; }
                             }
                         }
                     }
